+++ flake8_mock_spec.py
@@ -47,7 +47,8 @@ MOCK_MSG_LOOKUP = {
 
 # The attribute actually does exist, mypy reports that it doesn't
 PATCH_FUNCTION: str = mock.patch.__name__  # type: ignore
-PATCH_ARGS = frozenset(("new", "spec", "spec_set", "autospec", "new_callable"))
+PATCH_MULTIPLE_ARGS = frozenset(("spec", "spec_set", "autospec", "new_callable"))
+PATCH_ARGS = frozenset(("new", *PATCH_MULTIPLE_ARGS))
 PATCH_MSG_BASE = (
     f"%s unittest.mock.%s should be called with any of the {', '.join(PATCH_ARGS)} arguments, "
     f"{MORE_INFO_BASE}#fix-%s"
@@ -55,13 +56,32 @@ PATCH_MSG_BASE = (
 PATCH_CODE = f"{ERROR_CODE_PREFIX}020"
 PATCH_MSG = PATCH_MSG_BASE % (PATCH_CODE, PATCH_FUNCTION, PATCH_CODE.lower())
 PATCH_OBJECT_CODE = f"{ERROR_CODE_PREFIX}021"
-PATCH_OBJECT_FUNCTION = (PATCH_FUNCTION, "object")
+PATCH_OBJECT_FUNCTION = (PATCH_FUNCTION, mock.patch.object.__name__.rsplit("_", maxsplit=1)[-1])
 PATCH_OBJECT_MSG = PATCH_MSG_BASE % (
     PATCH_OBJECT_CODE,
     ".".join(PATCH_OBJECT_FUNCTION),
     PATCH_OBJECT_CODE.lower(),
 )
-PATCH_FUNCTIONS = frozenset((PATCH_FUNCTION, PATCH_OBJECT_FUNCTION))
+PATCH_MULTIPLE_FUNCTION = (
+    PATCH_FUNCTION,
+    mock.patch.multiple.__name__.rsplit("_", maxsplit=1)[-1],
+)
+PATCH_MULTIPLE_CODE = f"{ERROR_CODE_PREFIX}022"
+PATCH_MULTIPLE_MSG = PATCH_MSG_BASE % (
+    PATCH_MULTIPLE_CODE,
+    ".".join(PATCH_MULTIPLE_FUNCTION),
+    PATCH_MULTIPLE_CODE.lower(),
+)
+PATCH_ARGS_LOOKUP = {
+    PATCH_FUNCTION: PATCH_ARGS,
+    PATCH_OBJECT_FUNCTION: PATCH_ARGS,
+    PATCH_MULTIPLE_FUNCTION: PATCH_MULTIPLE_ARGS,
+}
+PATCH_MSG_LOOKUP = {
+    PATCH_FUNCTION: PATCH_MSG,
+    PATCH_OBJECT_FUNCTION: PATCH_OBJECT_MSG,
+    PATCH_MULTIPLE_FUNCTION: PATCH_MULTIPLE_MSG,
+}
 
 
 class Problem(NamedTuple):
@@ -78,15 +98,14 @@ class Problem(NamedTuple):
     msg: str
 
 
-def _get_fully_qualified_name(node: ast.expr) -> tuple[str, ...] | None:
+def _get_fully_qualified_name(node: ast.expr) -> tuple[str, ...]:
     """Retrieve the fully qualified name of a call func node.
 
     Args:
         node: The node to get the name of.
 
     Returns:
-        Tuple containing all the elements of the fully qualified name of the node or None if
-        unexpected nodes are found.
+        Tuple containing all the elements of the fully qualified name of the node.
     """
     if isinstance(node, ast.Name):
         return (node.id,)
@@ -94,23 +113,7 @@ def _get_fully_qualified_name(node: ast.expr) -> tuple[str, ...] | None:
         fully_qualified_parent = _get_fully_qualified_name(node.value)
         if fully_qualified_parent:
             return (*fully_qualified_parent, node.attr)
-    return None
-
-
-def _check_patch_keywords(node: ast.Call, msg: str) -> Problem | None:
-    """Check if the given patch call has expected arguments.
-
-    Args:
-        node: The patch call node to check.
-        msg: The error message to return if the check fails.
-
-    Returns:
-        Problem: If the patch call does not have the expected arguments.
-        None: If the patch call has the expected arguments.
-    """
-    if not any(keyword.arg in PATCH_ARGS for keyword in node.keywords):
-        return Problem(lineno=node.lineno, col_offset=node.col_offset, msg=msg)
-    return None
+    return ()
 
 
 class Visitor(ast.NodeVisitor):
@@ -147,16 +150,21 @@ class Visitor(ast.NodeVisitor):
                     )
                 )
 
-        if (
-            name == PATCH_FUNCTION
-            or fully_qualified_name is not None
-            and fully_qualified_name[-2:] == PATCH_OBJECT_FUNCTION
-        ):
-            problem = _check_patch_keywords(
-                node=node, msg=PATCH_MSG if name == PATCH_FUNCTION else PATCH_OBJECT_MSG
-            )
-            if problem:
-                self.problems.append(problem)
+        patch_msg_lookup_key = next(
+            (key for key in (name, fully_qualified_name[-2:]) if key in PATCH_MSG_LOOKUP),
+            None,
+        )
+        if patch_msg_lookup_key in PATCH_MSG_LOOKUP:
+            if not any(
+                keyword.arg in PATCH_ARGS_LOOKUP[patch_msg_lookup_key] for keyword in node.keywords
+            ):
+                self.problems.append(
+                    Problem(
+                        lineno=node.lineno,
+                        col_offset=node.col_offset,
+                        msg=PATCH_MSG_LOOKUP[patch_msg_lookup_key],
+                    )
+                )
 
         # Ensure recursion continues
         self.generic_visit(node)